allow blocklist built-in plugins
