+++ rundeckapp/grails-app/services/rundeck/services/ExecutionUtilService.groovy
@@ -204,6 +204,7 @@ class ExecutionUtilService {
                 } else {
                     args = new String[0];
                 }
+                //TODO: create a script node step item
                 return ExecutionItemFactory.createScriptFileItem(
                         cmd.getScriptInterpreter(),
                         cmd.getFileExtension(),
@@ -239,6 +240,7 @@ class ExecutionUtilService {
                             !!cmd.expandTokenInScriptFile
                     )
                 }else {
+                    //TODO: create a script-file node step item
                     return ExecutionItemFactory.createScriptFileItem(
                             cmd.getScriptInterpreter(),
                             cmd.getFileExtension(),
